Fixed bug where Reveal button for Picture Matching displays wrong order
The former method of swapping pictures in the Picture Matching game
involves swapping the Div containing the picture. This results to the
scrambling of the supposedly ordered id of the Divs. The solution was to
simply swap the backgroundImage attribute instead of the whole Div, to
maintain the order of the id.