Fixed background Gamma Correction
The green mountain background was gamma-corrected by Photoshop, resulting
to it looking mismatched even if they were using the same Hex color codes.
This was fixed using PNG Crush.

Also, loading of Google's Web Font was changed from http:// to only // to
allow it to be loaded over https also.